# ENH: speed up wide DataFrame.line plots by using a single LineCollection